Better post creation (#5)
* update `post` model

* refactor: `editor` and add new `create` `edit`...
...components

Putting the logic of creating and editing `Post` in one component was
not a good idea and it really made a mess. This commit fixes that by
removing existing `create-post` module.

It also adds new (empty) `create` and `edit` components.

* feat: Allow user to create new post using `create`

* feat: Allow user to edit post

- Add `edit.component` logic
- Remove formControl binding from tinymce
- Add `onChange` event listener on tinymce

* fix: Remove `form.dirty` validation in...
...valueChanges subscription

Featured image input change wasn't setting form as dirty and the post
was not being submitted because of this.
This commit fixes the issue by removing `isFormDirty()` validation

* feat: Add update request to `posts.service`

`create` and `update` routes were separated in API.
This commit updates the app according to backend API changes.